[New] Added test cases for transitions dialog
